requirements.txt
alembic==1.8.1
ascii_magic==2.3.0
git+https://github.com/abrichr/atomacos.git
bokeh==2.4.3
clipboard==0.0.4
deepdiff[optimize]==6.3.0
dictalchemy3==1.0.0
fire==0.4.0
ipdb==0.13.11
loguru==0.6.0
matplotlib==3.6.2
mss==6.1.0
openai==0.27.5
pandas==2.0.0
pygetwindow==0.0.9; sys_platform == 'win32'
pyinstaller
python-dotenv==1.0.0
pywin32==306; sys_platform == 'win32'
git+https://github.com/abrichr/pynput.git
pytest==7.1.3
rapidocr-onnxruntime==1.2.3
scikit-learn==1.2.2
scipy==1.9.3
sqlalchemy==1.4.43
tiktoken==0.4.0
torch==2.0.0
tqdm==4.64.0
transformers==4.28.1
